pyproject.toml
[tool.poetry]
name = "tellmewhattodo"
version = "3.0.0"
description = "Simple app that checks for what to do and presents that in a website"
authors = ["Martin Morset"]
license = "Apache 2.0"
packages = [{ include = "tellmewhattodo" }]

[tool.poetry.dependencies]
python = "^3.11"

pydantic = "^2.9"
requests = "^2.32"
click = "^8.1"
PyYAML = "^6.0"
pydantic-settings = "^2.6"
fastapi = {extras = ["standard"], version = "^0.115.4"}
sqlalchemy = "^2.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
black = "^24.10.0"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tools.black]
line-length = 100

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["FA102", "D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107"]